feat: integrate AI for card creation, enhance documentation, and improve localization
