+++ docs/deuda-tecnica.md
@@ -1,37 +1,51 @@
 # 🛠️ Registro de Deuda Técnica - TCG Nexus
 
-Este documento centraliza las mejoras pendientes, refactorizaciones y "magic strings" identificados que deben ser resueltos para mantener la calidad y escalabilidad del proyecto.
+## 1. Arquitectura y DX (Developer Experience)
 
-## 1. Magic Strings y Constantes de Dominio
+### ✅ Documentación de Servicios (Completado)
+*   **Acción:** Se implementó JSDoc y `@typedef` en `cardService.js` y `favoritesService.js`.
+*   **Resultado:** Autocompletado inteligente y mejor documentación de la estructura de datos.
 
-### Hallazgo: Opciones de Filtrado en `Home.jsx`
+### ✅ Refactor de "Fat Components" (Completado)
+*   **Acción:** Se extrajo la lógica de `Home.jsx` al hook personalizado `useInfiniteCards.js`.
+*   **Resultado:** Separación clara de incumbencias (Separation of Concerns) y componente Home puramente representativo.
+
+### ⏳ Constantes de Dominio
 *   **Archivo:** `src/pages/Home.jsx`
-*   **Problema:** Las constantes `TYPE_OPTIONS` y `RARITY_OPTIONS` están definidas localmente en la página.
-*   **Impacto:** Si se necesitan estos valores en la vista de detalle o en el creador de cartas por IA (US8), habrá duplicación de código.
-*   **Tarea Pendiente:** Mover a `src/config/constants.js` o `src/constants/domain.js`.
+*   **Problema:** `TYPE_OPTIONS` y `RARITY_OPTIONS` siguen definidos localmente.
+*   **Tarea Pendiente:** Mover a un archivo central (ej: `src/constants/game.js`) para uso compartido con la futura US8 (Creador de cartas).
 
-### Hallazgo: Mensajes de Interfaz Hardcodeados
-*   **Archivo:** `src/pages/Home.jsx`
-*   **Problema:** Mensajes como `"Invocando criaturas del Nexo..."`, `"Buscando..."` y mensajes de error están escritos directamente en el JSX o en el estado.
-*   **Tarea Pendiente:** Migrar estos strings a los archivos de traducción (`es.json` / `en.json`) durante la implementación de la **US2 (i18n)**.
+### 🚀 Escalabilidad del Estado
+*   **Contexto:** El sistema de favoritos usa `useState` y prop drilling/contexto simple.
+*   **Tarea Pendiente:** Evaluar el uso de un Store (Zustand o Redux Toolkit) si la complejidad del usuario (mazos, inventario) crece en la Parte 3 del TP.
 
-## 2. Arquitectura de Servicios
+## 2. Internacionalización y UI (Completado ✅)
 
-### Hallazgo: Centralización de Preferencias (Completado ✅)
-*   **Acción:** Se creó `preferencesService.js` para abstraer el uso de `localStorage` y `sessionStorage`.
-*   **Nota:** Mantener este patrón para cualquier nueva configuración de UI que se agregue (ej: temas, filtros persistentes).
+### ✅ Mensajes y Metadatos Localizados
+*   **Acción:** Se sincronizaron las claves entre `es.json` y `en.json`.
+*   **Acción:** Se implementó la actualización dinámica de `document.title` en `App.jsx`.
+*   **Acción:** Localizados tooltips y aria-labels en `ThemeToggle.jsx` y `Header.jsx`.
 
-## 3. Próximos Refactors Sugeridos
-*   **Validación de IDs:** En `cardService.js`, mejorar la validación de los IDs antes de realizar el fetch.
-*   **Tipado (JSDoc):** Agregar documentación de tipos a los retornos de los servicios para mejorar el autocompletado en los componentes.
+## 3. PWA y Despliegue
 
-## 4. Optimización de Repositorio (Git)
+### ✅ Configuración Base PWA (Completado)
+*   **Acción:** Creación de `manifest.json` y vinculación en `index.html`.
+*   **Tarea Pendiente:** Implementar un Service Worker (vía Vite PWA Plugin) para soporte offline real de las cartas ya cacheadas.
 
-### Hallazgo: Historial Pesado por Assets Originales
-*   **Problema:** Se subieron 78MB de imágenes Splash originales a la rama `develop` antes de implementar la estrategia de optimización.
-*   **Impacto:** El historial de Git retiene ese peso, haciendo que los clones del repositorio sean más lentos.
-*   **Decisión Temporal:** Se pospone la reescritura del historial (usando herramientas como `git filter-repo` o `BFG`) para evitar conflictos de merge críticos con el equipo durante el desarrollo del TP.
-*   **Tarea Pendiente:** Evaluar la limpieza profunda del historial antes de la entrega final o al cierre del cuatrimestre.
+### ⏳ Manejo de Entorno (Vite)
+*   **Tarea Pendiente:** Crear un `.env.example` para guiar la configuración de futuras integración (IA/Cloudinary).
+
+## 4. Calidad y Testing
+
+### ⏳ Cobertura de Tests
+*   **Estado:** Crítico/Pendiente.
+*   **Tarea Pendiente:** Configurar Vitest y React Testing Library. Priorizar testing de hooks personalizados (`useInfiniteCards`) y servicios.
+
+## 5. Optimización de Repositorio (Git)
+
+### ⏳ Historial Pesado por Assets Originales
+*   **Problema:** El repo retiene ~80MB de assets no optimizados en el historial.
+*   **Tarea Pendiente:** Ejecutar limpieza con `BFG Repo-Cleaner` antes del cierre del cuatrimestre.
 
 ---
-*Última actualización: Abril 2026*
+*Última actualización: Abril 2026 - Sesión de Cierre Tech Lead/QA*
