+++ src/components/ThemeToggle.jsx
@@ -1,7 +1,9 @@
 import { useState, useEffect } from 'react';
+import { useTranslation } from 'react-i18next';
 import { preferencesService } from '../services/preferencesService';
 
 export default function ThemeToggle() {
+  const { t } = useTranslation();
   const [theme, setTheme] = useState(() => preferencesService.getTheme());
 
   // Aplicar el tema al cargar y cuando cambie el estado
@@ -23,8 +25,8 @@ export default function ThemeToggle() {
     <button
       onClick={toggleTheme}
       className="flex h-10 w-10 items-center justify-center rounded-full border border-slate-300 bg-slate-100 text-lg shadow-lg transition-all hover:scale-110 hover:border-yellow-500 active:scale-95 dark:border-slate-700 dark:bg-slate-800 dark:hover:border-blue-400"
-      title={theme === 'dark' ? 'Invocar Modo Luz' : 'Abrazar las Sombras'}
-      aria-label="Cambiar tema"
+      title={theme === 'dark' ? t('theme.toLight') : t('theme.toDark')}
+      aria-label={t('theme.toggle')}
     >
       <span className="transform transition-transform duration-500 hover:rotate-12">
         {theme === 'dark' ? '🌙' : '☀️'}
