+++ src/pages/Home.jsx
@@ -1,9 +1,8 @@
-import { useState, useEffect, useCallback, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
-import cardService from '../services/cardService';
 import Card from '../components/Card';
 import SearchBar from '../components/SearchBar';
 import LoadingSpinner from '../components/LoadingSpinner';
+import { useInfiniteCards } from '../hooks/useInfiniteCards';
 
 export default function Home() {
   const { t, i18n } = useTranslation();
@@ -23,102 +22,25 @@ export default function Home() {
     t('home.filters.rarities.Legendario')
   ];
 
-  const [cards, setCards] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState(null);
-  const [activeFilters, setActiveFilters] = useState({ 
-    searchTerm: '', 
-    selectedTypes: [], 
-    selectedRarities: [] 
-  });
-
-  // Estados para Scroll Infinito
-  const [page, setPage] = useState(1);
-  const [hasMore, setHasMore] = useState(true);
-  const observer = useRef();
-  const LIMIT = 12;
-
-  // Elemento centinela para el IntersectionObserver
-  const lastCardElementRef = useCallback(node => {
-    if (isLoading) return;
-    if (observer.current) observer.current.disconnect();
-    
-    observer.current = new IntersectionObserver(entries => {
-      if (entries[0].isIntersecting && hasMore) {
-        setPage(prevPage => prevPage + 1);
-      }
-    });
-    
-    if (node) observer.current.observe(node);
-  }, [isLoading, hasMore]);
-
-  const fetchCards = useCallback(async (currentPage, filters, isNewSearch = false) => {
-    try {
-      setIsLoading(true);
-      setError(null);
-
-      const params = {
-        page: currentPage,
-        limit: LIMIT,
-      };
-
-      if (filters.searchTerm) {
-        params.search = filters.searchTerm;
-      }
-
-      const data = await cardService.getCards(params);
-
-      setCards(prevCards => {
-        if (isNewSearch) return data;
-        // Evitar duplicados por ID
-        const newCards = data.filter(newCard => !prevCards.some(pc => pc.id === newCard.id));
-        return [...prevCards, ...newCards];
-      });
-
-      setHasMore(data.length === LIMIT);
-    } catch (err) {
-      setError(t('home.error'));
-      console.error(err);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [t]);
-
-  // Efecto para carga inicial y cambios de página
-  useEffect(() => {
-    const load = async () => {
-      await fetchCards(page, activeFilters);
-    };
-    load();
-  }, [page, activeFilters, fetchCards]);
-
-  // Manejo de búsqueda y filtros
-  const handleSearch = useCallback(({ searchTerm, selectedTypes, selectedRarities }) => {
-    const newFilters = { searchTerm, selectedTypes, selectedRarities };
-    setActiveFilters(newFilters);
-    
-    setPage(1);
-    setHasMore(true);
-    fetchCards(1, newFilters, true);
-  }, [fetchCards]);
-
-  // Filtrado local para tipos y rarezas
-  const filteredCards = cards.filter(card => {
-    const cardData = card[lang] || card['es'];
-    
-    const matchType = activeFilters.selectedTypes.length === 0 || 
-                      activeFilters.selectedTypes.includes(cardData.type);
-    const matchRarity = activeFilters.selectedRarities.length === 0 || 
-                        activeFilters.selectedRarities.includes(cardData.rarity);
-    
-    return matchType && matchRarity;
+  const {
+    cards,
+    isLoading,
+    error,
+    hasMore,
+    page,
+    handleSearch,
+    lastCardElementRef
+  } = useInfiniteCards({
+    limit: 12,
+    initialFilters: { searchTerm: '', selectedTypes: [], selectedRarities: [] },
+    lang
   });
 
   if (error) {
     return (
       <div className="py-12 text-center">
         <div className="inline-block max-w-md rounded-xl border border-red-500/50 bg-red-500/10 p-6">
-          <p className="mb-4 font-medium text-red-400">{error}</p>
+          <p className="mb-4 font-medium text-red-400">{t('home.error')}</p>
           <button
             onClick={() => window.location.reload()}
             className="rounded-lg bg-red-500 px-6 py-2 text-white transition-colors hover:bg-red-600"
@@ -153,14 +75,14 @@ export default function Home() {
         </div>
       ) : (
         <>
-          {filteredCards.length === 0 ? (
+          {cards.length === 0 ? (
             <div className="py-12 text-center text-xl text-slate-400 dark:text-slate-500">
               {t('home.noResults')}
             </div>
           ) : (
             <div className={`grid grid-cols-1 gap-6 transition-opacity sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 ${isLoading && page === 1 ? 'opacity-50' : 'opacity-100'}`}>
-              {filteredCards.map((card, index) => {
-                const isLast = filteredCards.length === index + 1;
+              {cards.map((card, index) => {
+                const isLast = cards.length === index + 1;
                 return <Card key={card.id} card={card} ref={isLast ? lastCardElementRef : null} />;
               })}
             </div>
@@ -174,9 +96,8 @@ export default function Home() {
           )}
 
           {/* Mensaje de final del catálogo */}
-          {!hasMore && filteredCards.length > 0 && (
+          {!hasMore && cards.length > 0 && (
             <div className="mt-12 border-t border-slate-200 py-12 text-center text-slate-500 italic dark:border-slate-800">
-              {/* Aquí podrías agregar una clave 'home.allLoaded' si quieres */}
               Has llegado al final del Nexo.
             </div>
           )}
